cpp tests: Split Files
Split Code-Files for Unittests + Add kvs_check_hash tests
